+++ src/lib/menuActions.ts
@@ -1,0 +1,17 @@
+import { getModalStore } from '@skeletonlabs/skeleton'
+import type { LayoutNodeProxy } from './classes/layoutNodes'
+const modalStore = getModalStore()
+
+export const setParentID = (node: LayoutNodeProxy) => {
+    modalStore.trigger({
+        type: 'component',
+        component: 'fullNodeList',
+        meta: {
+            onClick: (e: CustomEvent) => {
+                node.setColumn("parent_node_id", e.detail.id)
+                modalStore.close()
+            },
+            exclude: [node.id, node.parent_node_id]
+        }
+    })
+}
